requirements.txt
cvxpy==1.2.1
matplotlib==3.2.1
natsort==8.2.0
numpy==1.18.2
opencv_python==4.6.0.66
Pillow==9.2.0
psutil==5.9.3
pycocotools==2.0.5
pynvml==11.4.1
rich==12.6.0
scikit_image==0.16.2
scikit_learn==1.1.2
scipy==1.4.1
skimage==0.0
tensorboard==2.1.1
tensorboardX==2.5.1
torch==1.4.0
torchvision==0.5.0
tqdm==4.64.1
yolox==0.3.0
